% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.gpar.R
\name{subset.gpar}
\alias{subset.gpar}
\title{Subsetting Grid Graphical Parameters}
\usage{
\method{subset}{gpar}(x, subset, ...)
}
\arguments{
\item{x}{a list of class \code{gpar}.}

\item{subset}{logical expression indicating elements to keep; missing values are taken as false.}

\item{...}{ignored.}
}
\value{
a list of class \code{gpar}.
}
\description{
Subsetting Grid Graphical Parameters
}
\details{
This is an S3 method for the generic \code{\link[base:subset]{base::subset()}} function, to subset graphical parameters that are vectors in a \code{gpar} object (a list typically produced by \code{\link[grid:gpar]{grid::gpar()}}).
}
\examples{
gp <- grid::gpar(pch = 21, fill = c("red", "blue", "green"))
gp13 <- subset(gp, c(TRUE, FALSE, TRUE))
print(gp13)

}
