% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_frame.R
\name{ternary_frame}
\alias{ternary_frame}
\title{Make a Triangular Frame for a Ternary Plot}
\usage{
ternary_frame(gp = grid::gpar(), ...)
}
\arguments{
\item{gp}{graphical settings via \code{\link[grid:gpar]{grid::gpar()}}}

\item{...}{arguments to pass to \code{\link[grid:grid.polygon]{grid::grid.polygon()}}}
}
\value{
A polygon grob object.
}
\description{
When drawn in a square viewport with x and y coordinates in the same units, the result is an equilateral triangle with one of its vertices at (0, 0).
}
\details{
The viewport should be square when the grob gets drawn.  This is not checked.
}
\examples{
tplot_new(corner = c(1, 1))
grid::grid.draw(ternary_frame())
#  Now draw another one, scaled down by factor 1 / 1.2
grid::pushViewport(ternary_viewport(corner = c(1, 1), scale = 1.2))
grid::grid.draw(ternary_frame(gp = grid::gpar(lty = "dashed")))

}
\author{
David Firth
}
