% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_lines.R
\name{ternary_lines}
\alias{ternary_lines}
\title{Draw lines in a Ternary Plot}
\usage{
ternary_lines(nby3, neg_action = "warning", ...)
}
\arguments{
\item{nby3}{a 3-column numeric matrix.  Each row represents one point to be plotted.  Rows will be scaled to sum to 1 before plotting, so no row can contain only zeros.}

\item{neg_action}{a character string, either \code{"warning"} or \code{"error"}; or \code{NULL} if no checking of negative values is wanted.  This provides optional checking for negative values in the \code{nby3} matrix.}

\item{...}{arguments to pass to \code{\link[grid:grid.lines]{grid::grid.lines()}}.}
}
\value{
A lines grob object.
}
\description{
Three-part compositional data points are connected by lines in a ternary plot.
}
\details{
When the resulting object is drawn, the viewport should be square.  This is not checked.
}
\examples{
endpoints <- matrix(c(-0.1, 0.3, 0.8,  ## this point will trigger a warning, by default 
                       0.2, 0.2, 0.6,
                       0.1, 0.8, 0.1),
                     3, 3, byrow = TRUE)
tplot_new()
grid::grid.draw(ternary_frame())
grid::grid.draw(ternary_lines(endpoints))
                      

}
\author{
David Firth
}
