% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_points.R
\name{ternary_points}
\alias{ternary_points}
\title{Make a Points Grob for a Ternary Plot}
\usage{
ternary_points(nby3, neg_action = "warning", gp = grid::gpar(), ...)
}
\arguments{
\item{nby3}{a 3-column numeric matrix.  Each row represents one point to be plotted.  Rows will be scaled to sum to 1 before plotting, so no row can contain only zeros.}

\item{neg_action}{a character string, either \code{"warning"} or \code{"error"}; or \code{NULL} if no checking of negative values is wanted.  This provides optional checking for negative values in the \code{nby3} matrix.}

\item{gp}{graphical settings via \code{\link[grid:gpar]{grid::gpar()}}}

\item{...}{arguments to pass to \code{\link[grid:grid.points]{grid::grid.points()}}.}
}
\value{
A points grob object (invisibly).
}
\description{
Three-part compositional data are converted to 2-dimensional coordinates and then displayed in a ternary plot.
}
\details{
When the resulting object is drawn, the viewport should be square.  This is not checked.
}
\examples{
## Toy data with compositions (1/6, 2/6, 3/6) and (2/5, 2/5, 1/5)
tplot_new()
grid::grid.draw(ternary_thirds())
toy_data <- matrix(c(1, 2, 3,
                     2, 2, 1),
                     2, 3, byrow = TRUE)
grid::grid.draw(ternary_points(toy_data, gp = grid::gpar(col = c("red", "blue"))))
grid::grid.draw(ternary_text(label = c("(1, 2, 3)", "(2, 2, 1)"), nby3 = toy_data,
                       hjust = c(-0.3, 0.5), vjust = c(0.5, -1)))
#' 
}
\author{
David Firth
}
