% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_thirds.R
\name{ternary_thirds}
\alias{ternary_thirds}
\title{Divide the Background of a Ternary Plot into Thirds}
\usage{
ternary_thirds(type = "vertices", fill = NULL, gp = grid::gpar(), vp = NULL)
}
\arguments{
\item{type}{either \code{"vertices"} (the default) or \code{"edges"}.}

\item{fill}{either \code{NULL} or a vector of 3 colours}

\item{gp}{a list of other graphical parameters, made by the \code{\link[grid:gpar]{grid::gpar()}} function.}

\item{vp}{a viewport to be used when drawing the resulting grob.}
}
\value{
A gList containing three polygon grobs.
}
\description{
The \code{\link[grid:grid.polygon]{grid::grid.polygon()}} constructor is used to create a background, which can be coloured, for a ternary plot.  The background divides the ternary plot into thirds.
}
\details{
A grob is constructed, to divide the equilateral triangle of a ternary plot into thirds.  Such a grob should normally be drawn \emph{before} other objects such as those that result from \code{\link[=ternary_frame]{ternary_frame()}} or \code{\link[=ternary_points]{ternary_points()}}, to serve as the background for the plot.

If the \code{type} is \code{"vertices"}, the thirds will be quadrilaterals into the three corners of the ternary plot.  If the \code{type} is \code{"edges"}, each third will be a triangle containing one edge and the centroid.  See examples below.

The current viewport should be square when the grob gets drawn.  This is not checked.
}
\examples{

## Coloured quadrilaterals into the 3 vertices
tv <- ternary_thirds(type = "vertices",
                     fill = c("#ffcccc", "#ccffcc", "#ccccff"))
tplot_new(); grid::grid.draw(tv)

## Coloured triangles along the 3 edges, here with zero-width frame lines
te <- ternary_thirds(type = "edges",
                     fill = c("cyan3", "magenta3", "yellow3"),
                     gp = grid::gpar(col = "white"))
tplot_new(); grid::grid.draw(te)

}
\author{
David Firth
}
