% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_xy.R
\name{ternary_xy}
\alias{ternary_xy}
\title{Convert Three-Part Composition Vectors to Coordinates for Plotting}
\usage{
ternary_xy(nby3, neg_action = "warning")
}
\arguments{
\item{nby3}{a three-column numeric matrix.  Each row represents one point to be plotted.  Rows will be scaled to sum to 1 before plotting, so no row can contain only zeros.}

\item{neg_action}{a character string, either \code{"warning"} or \code{"error"}; or \code{NULL} if no checking of negative values is wanted.  This provides optional checking for negative values in the \code{compos} matrix.}
}
\value{
A two-component list of \code{"unit"} objects, with component names \code{"x"} and \code{"y"}.
}
\description{
Converts a 3-column matrix of compositional data to a 2-component list of x and y coordinates for ternary plotting in the native units of the (square) viewport.
}
\details{
This is an internal helper function, to enable the making/drawing of ternary plot objects in square viewports.  It is called by functions such as \code{\link[=ternary_frame]{ternary_frame()}} and \code{\link[=ternary_points]{ternary_points()}}.  This function is invariant to scaling of rows in the data, and optionally tolerates negative values in the data.
}
\examples{
corners_and_centre <- rbind(diag(3), rep(1/3, 3))
ternary_xy(corners_and_centre)

}
\author{
David Firth
}
