% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tplot_new.R
\name{tplot_new}
\alias{tplot_new}
\title{Start a New Ternary Plot or Tetraplot}
\usage{
tplot_new(...)
}
\arguments{
\item{...}{arguments to be passed to \code{\link[=ternary_viewport]{ternary_viewport()}}.}
}
\value{
A viewport object (invisibly).
}
\description{
This function erases the current device or moves to a new page, and then sets up a square viewport suitable for drawing a ternary plot or a tetraplot.  See \code{\link[=ternary_viewport]{ternary_viewport()}} for details such as how to set the orientation and/or scale.
}
\details{
This is just a convenient wrapper for a common sequence: a call to \code{\link[grid:grid.newpage]{grid::grid.newpage()}} and then \code{\link[=ternary_viewport]{ternary_viewport()}} and then \code{\link[grid:viewports]{grid::pushViewport()}}.
}
\examples{
## Draw a ternary frame, oriented towards top left and smaller than the viewport
tplot_new(corner = c(0, 1), scale = 1.2)
grid::grid.rect(gp = grid::gpar(col = "green", lwd = 10))
# The viewport boundary was drawn in green.
grid::grid.draw(ternary_frame())

}
\author{
David Firth
}
